[Signature Pad] Converted to RelativeLayout.
SignaturePadView now inherits from RelativeLayout, which allows us to
set the position of
children views in a parametric way, using the RelativeLayout engine.
This way we no longer
need to override OnLayout with an arbitrary algorythm with fixed
positioning policies, and gives
the developer more control over the positioning of the elements for the
SignaturePad.