[SignaturePad] Android: Don't store strokes in a bitmap by default (#55016)
Storing this bitmap raises the memory requirements for the widget too
high. Developers with heavy applications may have memory allowance
exceptions when trying to use the SignaturePad. So now by default, we
redraw all of the paths whenever the widget is invalidated. It seems to
respond pretty well, but I'm leaving the old behavior available as an
option to the developer as an attribute. iOS version not reviewed until
we get any complaints.