Making an optimization in the Signature Pad View
Found that the GetImage methods were spending most of the time in the
Points property building the array.

The Points property has been improved to not built several temporary arrays.

getCroppedRectangle has been changed to only call Points once and reuse the
result.

IsBlank property should not have to calculate the entire array, only find one
point and return false.

This fix is in response to case 121796